+++ core/apps.js
@@ -2,10 +2,13 @@ import { AboutApp } from "../apps/about/index";
 import { ChangelogApp } from "../apps/changelog";
 import { CubeApp } from "../apps/cube";
 import { DialogApp } from "../apps/dialog";
+import { TerminalApp } from "../apps/terminal";
 
 export const apps = {
   about: AboutApp,
   changelog: ChangelogApp,
   cube: CubeApp,
+  terminal: TerminalApp,
+
   dialog: DialogApp
 }